+++ hackerrank/ex1.py
