+++ hackerrank/ex2.py
