+++ hackerrank/ex3.java
