+++ hackerrank/ex4.py
