--- potencia.py
@@ -1,11 +1,0 @@
-"""
-Fazer algoritmo que calcula o resultado de uma base e expoente dado.
-(Tentar sair da complexidade O(n) para uma menor, ou resolver de forma recursiva).
-Ou seja, calcular a potencia SEM usar a funcao nativa da linguagem ou a operacao de ^. Resolver utilizando
-apenas as 4 operações básicas e resto de divisão (caso seja necessário)
-
-input -> base: int
-         expoente: int
-
-output -> resultado: int
-"""
